inject jsenvtoolbar into source files as well
